+++ src/Watch/Fallback.php
@@ -3,24 +3,21 @@ declare(strict_types = 1);
 
 namespace Innmind\FileWatch\Watch;
 
-use Innmind\FileWatch\{
-    Watch,
-    Ping,
-};
+use Innmind\FileWatch\Ping;
 use Innmind\Url\Path;
 
-final class Fallback implements Watch
+/**
+ * @internal
+ */
+final class Fallback
 {
-    private Watch $attempt;
-    private Watch $fallback;
-
-    public function __construct(Watch $attempt, Watch $fallback)
-    {
-        $this->attempt = $attempt;
-        $this->fallback = $fallback;
+    public function __construct(
+        private Tailf $attempt,
+        private Stat $fallback,
+    ) {
     }
 
-    public function __invoke(Path $file): Ping
+    public function __invoke(Path $file): Ping\Implementation
     {
         return new Ping\Fallback(
             ($this->attempt)($file),