+++ src/Watch/Kind.php
@@ -3,24 +3,21 @@ declare(strict_types = 1);
 
 namespace Innmind\FileWatch\Watch;
 
-use Innmind\FileWatch\{
-    Watch,
-    Ping,
-};
+use Innmind\FileWatch\Ping;
 use Innmind\Url\Path;
 
-final class Kind implements Watch
+/**
+ * @internal
+ */
+final class Kind
 {
-    private Watch $files;
-    private Watch $directories;
-
-    public function __construct(Watch $files, Watch $directories)
-    {
-        $this->files = $files;
-        $this->directories = $directories;
+    public function __construct(
+        private Fallback $files,
+        private Stat $directories,
+    ) {
     }
 
-    public function __invoke(Path $file): Ping
+    public function __invoke(Path $file): Ping\Implementation
     {
         return match ($file->directory()) {
             true => ($this->directories)($file),