+++ src/Watch/Tailf.php
@@ -3,32 +3,31 @@ declare(strict_types = 1);
 
 namespace Innmind\FileWatch\Watch;
 
-use Innmind\FileWatch\{
-    Watch,
-    Ping,
-};
+use Innmind\FileWatch\Ping;
 use Innmind\Url\Path;
 use Innmind\Server\Control\Server\{
     Processes,
     Command,
 };
 
-final class Tailf implements Watch
+/**
+ * @internal
+ */
+final class Tailf
 {
-    private Processes $processes;
-
-    public function __construct(Processes $processes)
-    {
-        $this->processes = $processes;
+    public function __construct(
+        private Processes $processes,
+    ) {
     }
 
-    public function __invoke(Path $file): Ping
+    public function __invoke(Path $file): Ping\ProcessOutput
     {
         return new Ping\ProcessOutput(
             $this->processes,
             Command::foreground('tail')
                 ->withShortOption('f')
-                ->withArgument($file->toString()),
+                ->withArgument($file->toString())
+                ->streamOutput(),
         );
     }
 }